disallow string refs
string refs are considered legacy and the callback approach is the recommended way forward.
https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/no-string-refs.md

String refs will eventually go away in React core.
https://github.com/facebook/react/pull/4174
https://github.com/facebook/react/issues/6350
https://github.com/facebook/react/issues/6250

ReactCurrentOwner is the only remaining piece of shared global state in React internals. String refs are the primary reason it is still around.

The benefit of adopting this now (pre-deprecation) is to be ahead of the curve when that does happen.