Inject meta tags from iconstats.json
iconstats.json is generated by favicon-webpack-plugin and includes a
manifest.json file. It contains all of the required icon assets, so all
we have to add is add the /js path so they resolve correctly.

Remove sw.js file

This was the result of an incorrect path for the build and should not be
tracked