Create service worker registration module
This file will be required using require.ensure so that it compiles to
its own file, for caching. It exports a function that must be called to
trigger the registration. The function accepts a callback that will be
called when the registration object is available.

Initially this used promises, but that caused the file size to go up a
lot since webpack was included polyfills for Promise + friends.